requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.9.0
frozenlist==1.3.1
identify==2.5.15
idna==3.4
multidict==6.0.2
nodeenv==1.7.0
platformdirs==2.6.2
pre-commit==3.0.1
py-cord==2.3.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
ruff==0.0.236
urllib3==1.26.12
virtualenv==20.17.1
yarl==1.8.1
